update readme, move JsonDbCodec to json package
