+++ magnum-pg/src/main/scala/com/augustnagro/magnum/pg/json/JsonDbCodec.scala
@@ -1,4 +1,4 @@
-package com.augustnagro.magnum.pg
+package com.augustnagro.magnum.pg.json
 
 import com.augustnagro.magnum.DbCodec
 import org.postgresql.util.PGobject
